pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "netlanventory"
version = "0.1.0"
description = "Modular network scanning and inventory tool"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "NetLanVentory Contributors" }]

dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    # Configuration
    "pydantic-settings>=2.2.1",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    # CLI
    "click>=8.1.7",
    "rich>=13.7.1",
    # Networking / scanning
    "scapy>=2.5.0",
    "python-nmap>=0.7.1",
    # Logging
    "structlog>=24.1.0",
    # Async utilities
    "anyio>=4.3.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.4.4",
    "mypy>=1.10.0",
]

[project.scripts]
netlv = "netlanventory.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["netlanventory"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
